reformatted inconsistencies, separated all functions into their own modules, created test files for all modules
