+++ src/modules/countdown.js
@@ -1,20 +1,5 @@
 'use strict';
 
-// TO GO ON INDEX.JS
-
-// example:
-// user should type /countdown 00:00:10
-// to start a 10 second countdown
-app.command('/countdown', async ({ message, ack, say }) => {
-
-  await ack();
-  console.log(message);
-  let output = setCountdown(message.text);
-
-  await say(output);
-
-});
-
 
 const setCountdown = async ({ message, say }) => {
   // validate input pattern using regex
